+++ crates/circuit_library/src/qdrift.rs
@@ -10,15 +10,18 @@
 // copyright notice, and modified files need to carry a notice indicating
 // that they have been altered from the originals.
 
+use num_complex::Complex64;
 use qiskit_circuit::Qubit;
 use qiskit_circuit::circuit_data::{CircuitData, CircuitDataError};
 use qiskit_circuit::operations::{Param, StandardInstruction, multiply_param, radd_param};
 use qiskit_circuit::packed_instruction::PackedOperation;
 use qiskit_quantum_info::sparse_observable::{SparseObservable, SparseTermView};
-use qiskit_synthesis::evolution::trotterization::{
-    TrotterizationError, qdrift_evolution as qdrift, reorder_terms,
-};
+use qiskit_synthesis::evolution::suzuki_trotter::reorder_terms;
 use qiskit_synthesis::pauli_evolution::{Instruction, sparse_term_evolution};
+use rand::distr::weighted::{Error as DistroError, WeightedIndex};
+use rand::prelude::*;
+use rand::rngs::SysRng;
+use rand_pcg::Pcg64Mcg;
 use smallvec::smallvec;
 use thiserror::Error;
 
@@ -34,7 +37,7 @@ pub fn qdrift_evolution(
 
     let evo: Vec<(usize, f64)> = match qdrift(time, reps, seed, observable.coeffs().iter()) {
         Ok(evo) => evo,
-        Err(err) => return Err(EvolutionError::FailedEvolutionError(err)),
+        Err(err) => return Err(err),
     };
 
     let sampled_paulis = evo.iter().map(|(index, coeff)| {
@@ -108,6 +111,59 @@ pub fn qdrift_evolution(
     }
 }
 
+pub fn qdrift<'a>(
+    time: f64,
+    reps: u32,
+    seed: Option<u64>,
+    coeffs_iter: impl Iterator<Item = &'a Complex64>,
+) -> Result<Vec<(usize, f64)>, EvolutionError> {
+    let mut rng = match seed {
+        Some(seed) => Pcg64Mcg::seed_from_u64(seed),
+        None => Pcg64Mcg::try_from_rng(&mut SysRng).unwrap(),
+    };
+    let mut lambd = 0.0;
+
+    match coeffs_iter
+        .enumerate()
+        .map(|(i, coeff)| match real_or_fail(coeff) {
+            Ok(real_coeff) => Ok({
+                lambd += real_coeff;
+                (i, real_coeff)
+            }),
+            Err(err) => Err(err),
+        })
+        .collect::<Result<Vec<_>, _>>()
+    {
+        Ok(coeffs) => {
+            let num_gates = (2.0 * lambd.powi(2) * time.powi(2) * reps as f64).ceil() as usize;
+            match WeightedIndex::new(coeffs.iter().map(|(_, coeff)| coeff.abs() / lambd)) {
+                Ok(distr) => Ok({
+                    (0..num_gates)
+                        .map(|_| {
+                            let (index, coeff) = coeffs[distr.sample(&mut rng)];
+                            (
+                                index,
+                                coeff.signum() * (2.0 * lambd / (num_gates as f64) * time),
+                            )
+                        })
+                        .collect()
+                }),
+                Err(err) => Err(EvolutionError::DistributionError(err)),
+            }
+        }
+        Err(err) => Err(err),
+    }
+}
+
+/// Internal helper to extract real part of a complex number,
+/// returning an error if imaginary part is non-zero.
+fn real_or_fail(z: &Complex64) -> Result<f64, EvolutionError> {
+    if z.im.abs() > 1e-12 {
+        return Err(EvolutionError::RealOrFail(z.im.abs()));
+    }
+    Ok(z.re)
+}
+
 #[derive(Debug, Error)]
 pub enum EvolutionError {
     /// A general error for terms reordering error
@@ -119,5 +175,13 @@ pub enum EvolutionError {
     CircuitBuild(#[from] CircuitDataError),
 
     #[error["qDrift evolution failed"]]
-    FailedEvolutionError(#[from] TrotterizationError),
+    FailedEvolutionError(),
+
+    /// Complex value obtained from real approximation
+    #[error["Encountered complex value {0}, but expected real."]]
+    RealOrFail(f64),
+
+    /// Couldn't generate weighted distribution
+    #[error["Failed creating weight distribution"]]
+    DistributionError(#[from] DistroError),
 }
