+++ crates/synthesis/src/evolution/mod.rs
@@ -77,4 +77,4 @@ pub fn evolution(m: &Bound<PyModule>) -> PyResult<()> {
     Ok(())
 }
 
-pub mod trotterization;
+pub mod suzuki_trotter;
