+++ test/c/test_qdrift.c
@@ -42,7 +42,7 @@ static int test_qdrift_single_zz(void) {
     QkObsTerm term = {coeff, 2, bit_terms, indices, 2};
     qk_obs_add_term(obs, &term);
 
-    QkCircuit *circ = qk_qdrift(obs, 1, 0.5, 0, true, false);
+    QkCircuit *circ = qk_circuit_library_qdrift(obs, 1, 0.5, 0, true, false);
 
     size_t instr_count = qk_circuit_num_instructions(circ);
     if (instr_count != 1) {
@@ -104,7 +104,7 @@ static int test_qdrift_xi_plus_zz(void) {
     QkObsTerm term_2 = {coeff_2, 2, bit_term_2, indices_2, 2};
     qk_obs_add_term(obs, &term_2);
 
-    QkCircuit *circ = qk_qdrift(obs, 1, 0.5, 0, true, false);
+    QkCircuit *circ = qk_circuit_library_qdrift(obs, 1, 0.5, 0, true, false);
 
     size_t instr_count = qk_circuit_num_instructions(circ);
 
@@ -187,7 +187,7 @@ static int test_qdrift_gate_count_scaling(void) {
 
     /* Case A: t=0.5 → num_gates = 1 */
     {
-        QkCircuit *circ = qk_qdrift(obs, 1, 0.5, 0, true, false);
+        QkCircuit *circ = qk_circuit_library_qdrift(obs, 1, 0.5, 0, true, false);
 
         size_t n = qk_circuit_num_instructions(circ);
         if (n != 1) {
@@ -201,7 +201,7 @@ static int test_qdrift_gate_count_scaling(void) {
 
     /* Case B: t=1.0 → num_gates = 2 */
     {
-        QkCircuit *circ = qk_qdrift(obs, 1, 1.0, 0, true, false);
+        QkCircuit *circ = qk_circuit_library_qdrift(obs, 1, 1.0, 0, true, false);
 
         size_t n = qk_circuit_num_instructions(circ);
         if (n != 2) {
@@ -215,7 +215,7 @@ static int test_qdrift_gate_count_scaling(void) {
 
     /* Case C: reps=2, t=1.0 → num_gates = 4 */
     {
-        QkCircuit *circ = qk_qdrift(obs, 2, 1.0, 0, true, false);
+        QkCircuit *circ = qk_circuit_library_qdrift(obs, 2, 1.0, 0, true, false);
 
         size_t n = qk_circuit_num_instructions(circ);
         if (n != 4) {
@@ -249,7 +249,7 @@ static int test_qdrift_qubit_bounds(void) {
         qk_obs_add_term(obs, &term);
     }
 
-    QkCircuit *circ = qk_qdrift(obs, 1, 1.0, 0, true, false);
+    QkCircuit *circ = qk_circuit_library_qdrift(obs, 1, 1.0, 0, true, false);
 
     size_t n = qk_circuit_num_instructions(circ);
     for (size_t k = 0; k < n; ++k) {
@@ -291,7 +291,7 @@ static int test_qdrift_seed_auto_generation(void) {
     qk_obs_add_term(obs, &term3);
 
     // Let the seed be automatically generated by passing zero as seed value
-    QkCircuit *qc = qk_qdrift(obs, reps, time, 0, true, false);
+    QkCircuit *qc = qk_circuit_library_qdrift(obs, reps, time, 0, true, false);
 
     // Since the generation of gates is based on a weighted distribution,
     // if all the coeffs are the same, they will have the same weight,
@@ -358,7 +358,7 @@ static int test_qdrift_multiple_terms_manual(void) {
     QkObsTerm term2 = {(QkComplex64){1.0, 0.0}, 3, bits2, (uint32_t[3]){0, 1, 2}, num_qubits};
     qk_obs_add_term(obs, &term2);
 
-    QkCircuit *qc = qk_qdrift(obs, reps, time, 1, true, false);
+    QkCircuit *qc = qk_circuit_library_qdrift(obs, reps, time, 1, true, false);
 
     int result = (compare_circuits(qc, expected)) ? Ok : EqualityError;
 
@@ -388,7 +388,7 @@ static int test_qdrift_reorder(void) {
     QkObsTerm term2 = {(QkComplex64){1.0, 0.0}, 1, op2_bits, (uint32_t[1]){0}, 1};
     qk_obs_add_term(obs, &term2);
 
-    QkCircuit *qc = qk_qdrift(obs, 1, 0.65, 13, false, false);
+    QkCircuit *qc = qk_circuit_library_qdrift(obs, 1, 0.65, 13, false, false);
 
     if (!compare_circuits(qc, expected))
         result = EqualityError;
@@ -416,7 +416,7 @@ static int test_qdrift_reorder(void) {
     QkObsTerm term22 = {(QkComplex64){1.0, 0.0}, 1, op22_bits, (uint32_t[1]){0}, 1};
     qk_obs_add_term(obs2, &term22);
 
-    QkCircuit *qc2 = qk_qdrift(obs2, 1, 0.65, 13, true, false);
+    QkCircuit *qc2 = qk_circuit_library_qdrift(obs2, 1, 0.65, 13, true, false);
 
     if (!compare_circuits(qc2, expected2))
         result = EqualityError;
@@ -443,7 +443,7 @@ static int test_qdrift_insert_barrier(void) {
     QkObsTerm term2 = {(QkComplex64){1.0, 0.0}, 1, op2_bits, (uint32_t[1]){0}, 1};
     qk_obs_add_term(obs, &term2);
 
-    QkCircuit *qc = qk_qdrift(obs, 3, 0.238, 1, true, true);
+    QkCircuit *qc = qk_circuit_library_qdrift(obs, 3, 0.238, 1, true, true);
 
     int result = (compare_circuits(qc, expected)) ? Ok : EqualityError;
